visual_pizza_agent/backend/pyproject.toml
[project]
name = "visual_pizza_agent"
version = "0.1.0"
description = "A simple pizza agent using the Live API for a conversational experience"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "google-adk",
    "pillow>=12.1.1",
]

[tool.uv]
package = false
